requirements.txt
Django==3.0.3
djangorestframework==3.11.0
numpy==1.18.5
sklearn-pandas==1.8.0
traitlets==4.3.3
ujson==1.35
urllib3==1.25.9
scikit-learn==0.23.1
joblib==0.16.0
gunicorn==19.0.0